+++ src/lib/terra-components.module.ts
@@ -18,6 +18,9 @@ import { ModalModule } from 'ngx-bootstrap/modal';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { A11yModule } from '@angular/cdk/a11y';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatSelectModule } from '@angular/material/select';
+import { MatInputModule } from '@angular/material/input';
 
 @NgModule({
     imports: [
@@ -37,7 +40,10 @@ import { A11yModule } from '@angular/cdk/a11y';
         DragDropModule,
         MatIconModule,
         MatCheckboxModule,
-        A11yModule
+        A11yModule,
+        MatFormFieldModule,
+        MatSelectModule,
+        MatInputModule
     ],
     declarations: [...components, ...directives, ...pipes],
     exports: [...exportedComponents, ...exportedDirectives, ...pipes]